refactor: move util to own file
