+++ src/commands/open.ts
@@ -1,7 +1,8 @@
 import path from "path";
 import yargs from "yargs";
 import { PLACEFILE_NAME } from "../constants";
-import { getCommandName, getSettings } from "../util/getSettings";
+import { getSettings } from "../util/getSettings";
+import { getCommandName } from "../util/getCommandName";
 import { getWindowsPath } from "../util/getWindowsPath";
 import { identity } from "../util/identity";
 import { run } from "../util/run";
