+++ src/util/getSettings.ts
@@ -8,7 +8,3 @@ export async function getSettings(projectPath: string) {
 	const pkgJson = packageJsonType.parse(JSON.parse(pkgJsonContents));
 	return pkgJson?.["rbxts-build"] ?? {};
 }
-
-export function getCommandName(settings: packageJsonType["rbxts-build"], command: string) {
-	return settings?.names?.[command] ?? command;
-}
